Hosting endpoint (#4)
* add: add Hosting and init your logic

* add: continue hosting logic

* add: logic to phone and guest

+++ src/main/java/hotel/domain/repository/GuestRepository.java
@@ -1,6 +1,7 @@
 package hotel.domain.repository;
 
 import hotel.data.entity.guest.Guest;
+import hotel.data.entity.guest.Phone;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -8,7 +9,7 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Repository
-public interface GuestRepository extends JpaRepository<Guest, Long> {
+public interface GuestRepository extends JpaRepository<Guest, Integer> {
 
     Optional<Guest> findByRg(String rg);
     Optional<Guest> findByRgOrDocument(String rg, String document);
